+++ pyannote/audio/core/io.py
@@ -40,8 +40,6 @@ import torchaudio
 from pyannote.core import Segment
 from torch import Tensor
 
-torchaudio.set_audio_backend("soundfile")
-
 AudioFile = Union[Text, Path, IOBase, Mapping]
 
 AudioFileDocString = """